OP-1923 - abstracted dirmap for reuse in hosts
Implemented for Maya (refactored current implementation)
